// crematorium/src/types/types.ts
export type Crematorium = {
    name?: string;
    age?: string;
    approve_state?: string;
    canton?: string;
    community?: string;
    crematory?: string;
    date?: Date | null | undefined
    deathcertificate?: string | null | undefined
    deceased?: string | null | undefined
    district?: string;
    furnace?: string;
    houseregistration?: Date | null | undefined
    idcard?: Date | null | undefined
    leafnumber?: string | null;
    nameregister?: string | null;
    payment?: string | null;
    phone?: string | null;
    province?: string;
    relationship?: string | null;
    time?: string;
    transaction_date?: Date | null | undefined
    username?: string | null;
    village?: string | null;

}